pyproject.toml
[project]
name = "pdf-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "langchain-text-splitters>=0.3.8",
    "pymupdf>=1.26.3",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
